Move get_supported_features to helpers/entity.py, add tests
